Fixup ownership of the /tmp/downloads directory (#1)
It should be owned by the user that is set in the config file now.

It is actually possible to set a different user in the config file,
by means of using `docker run -v /etc/mailerq:/etc/mailerq`, as that
will cause the mailerq process in the docker container to use the
local config file. But in any case, this fixes it for the case that
the user is the default `mailerq` user.